+++ cashocs/_optimization/line_search/polynomial_line_search.py
@@ -81,7 +81,7 @@ class PolynomialLineSearch(line_search.LineSearch):
             satisfied.
 
         """
-        if solver.iteration >= solver.maximum_iterations:
+        if solver.iteration >= solver.max_iter:
             return True
 
         if self.stepsize * self.search_direction_inf <= 1e-8: